soma-279/refactor - refactoring api codes
