+++ api/preference.ts
@@ -1,28 +1,17 @@
+import { useRouter } from 'next/router';
+
+import { useMutation, UseMutationResult } from 'react-query';
+
 import { InfoState } from '#types/storeType/info';
+import { AxiosError } from 'axios';
 import { Axios } from 'lib/axios';
 import { arrToString } from 'utils';
-
-export type PreferenceRequest = {
-  styles: number[];
-  gender: string;
-  height: string;
-  bodyShape: string;
-  topSize: string[];
-  bottomSize: string[];
-  topColor: string[];
-  bottomColor: string[];
-};
-
-export type PreferenceResponse = {
-  staus: number;
-  style: string;
-};
 
 const removeBlank = (str: string) => str.replace(/ /g, '');
 
 export const postPreference = async (
   state: InfoState,
-): Promise<PreferenceResponse> => {
+): Promise<res.Preference> => {
   const { topSize, bottomSize, topColors, bottomColors } = state;
 
   const requestData = {
@@ -34,9 +23,27 @@ export const postPreference = async (
   };
   delete requestData.infoUpdate;
 
-  const { data } = await Axios.post<PreferenceResponse>(
+  const { data } = await Axios.post<res.Preference>(
     '/api/preference/save',
     JSON.stringify(requestData),
   );
   return data;
 };
+
+export function usePostPreference(): UseMutationResult<
+  res.Preference,
+  AxiosError,
+  InfoState
+> {
+  const router = useRouter();
+
+  return useMutation(postPreference, {
+    onSuccess: (data) => {
+      console.log(data);
+      router.push('/');
+    },
+    onError: (error) => {
+      console.log(error);
+    },
+  });
+}
