+++ hooks/index.ts
@@ -2,8 +2,6 @@ import useDebounceInput from './useDebounceInput';
 import useDidMountEffect from './useDidMountEffect';
 import useDragScroll from './useDragScroll';
 import useMounted from './useMounted';
-import usePostAuthToken from './usePostAuthToken';
-import usePostPreference from './usePostPreference';
 import useThrottleInput from './useThrottleInput';
 import useTimeForToday from './useTimeForToday';
 
@@ -12,8 +10,6 @@ export {
   useDidMountEffect,
   useDragScroll,
   useMounted,
-  usePostAuthToken,
-  usePostPreference,
   useThrottleInput,
   useTimeForToday,
 };
